pyproject.toml
[project]
name = "syftbox"
version = "0.2.8"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.114.0",
    "uvicorn>=0.30.6",
    "jinja2>=3.1.4",
    "typing-extensions>=4.12.2",
    "markdown>=3.7",
    "pandas>=2.2.2",
    "postmarker>=1.0",
    "watchdog>=5.0.2",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.5.2",
    "pytest>=8.3.3",
    "httpx>=0.27.2",
    "pyyaml>=6.0.2",
    "psutil>=6.0.0",
    "loguru>=0.7.2",
    "py-fast-rsync",
    "pathspec>=0.12.1",
    "python-multipart>=0.0.12",
    "rich>=13.9.2",
    "croniter>=3.0.0",
    "typer>=0.12.5",
    "pid>=3.0.4",
    "pydantic[email]>=2.9.2",
    "ruamel-yaml>=0.18.6",
    "pyjwt>=2.10.0",
]

[project.optional-dependencies]
# add using `uv add --optional <group> <pip package>`
# will be referenced in the built wheel


[project.scripts]
syftbox = "syftbox.main:main"

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.uv]
# add using `uv add --dev <group> <pip package>`
# this will be completely ignored in the built wheel
dev-dependencies = [
    "bump2version>=1.0.1",
    "faker>=30.3.0",
    "ipykernel>=6.29.5",
    "locust>=2.32.0",
    "pre-commit>=4.0.1",
    "pytest-cov>=5.0.0",
    "pytest-httpx>=0.33.0",
    "pytest-timeout>=2.3.1",
    "pytest-xdist[psutil]>=3.6.1",
    "pytest>=8.3.3",
]

[tool.setuptools]
packages = { find = {} }    # Use the find directive as a table
include-package-data = true # Include package data

[tool.setuptools.package-data]
syftbox = ["*.css", ".js", ".html", ".zip", ".sh"]

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.ruff]
line-length = 120
exclude = ["data", "users", "build", "dist", ".venv"]

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = ["F401"]

[tool.coverage.report]
skip_empty = true
